separation of concerns on the backend - transcription controller added

+++ app/backend/src/audio-transcription.js
@@ -1,6 +1,6 @@
-import { pipeline } from '@huggingface/transformers'
+import {pipeline} from '@huggingface/transformers'
 import fs from 'fs'
-import { decode } from 'wav-decoder'
+import {decode} from 'wav-decoder'
 import ffmpeg from 'fluent-ffmpeg'
 
 // Download the Whisper model (if not already downloaded)
@@ -35,7 +35,7 @@ function convertTo16KHzMono(inputFilePath, outputFilePath) {
 async function readWavFile(filePath) {
     const buffer = fs.readFileSync(filePath)
     const decodedWav = await decode(buffer)
-    return { sampleRate: decodedWav.sampleRate, channelData: decodedWav.channelData[0] }
+    return {sampleRate: decodedWav.sampleRate, channelData: decodedWav.channelData[0]}
 }
 
 /**
@@ -61,11 +61,13 @@ function sliceAudio(channelData, sampleRate, chunkLength) {
  * @returns {Promise<string>} Promise with the transcribed text
  */
 export async function transcribeAudio(filePath) {
-    let { sampleRate, channelData } = await readWavFile(filePath)
+    let {sampleRate, channelData} = await readWavFile(filePath)
 
-    let tempFilePath = '../backend/samples/temp.wav'
+    let tempFilePath = '../backend/samples/converted.wav'
+    let isConverted = false
     if (sampleRate !== 16000) {
         await convertTo16KHzMono(filePath, tempFilePath)
+        isConverted = true
 
         // update decoded data if the file was converted
         const decodedData = await readWavFile(tempFilePath)
@@ -78,9 +80,11 @@ export async function transcribeAudio(filePath) {
         transcriptionResult += result.text
     }
 
-    fs.unlink(tempFilePath, (err) => {
-        if (err) console.error('ERROR: Error deleting temp file:', err)
-    })
+    if (isConverted) {
+        fs.unlink(tempFilePath, (err) => {
+            if (err) console.error('ERROR: Error deleting temp file:', err)
+        })
+    }
 
     return transcriptionResult.trim()
 }
